+++ Tests/Runtime/FixturesTests.cs
@@ -11,6 +11,14 @@ public class FixturesTests {
     }
 
     [Test]
+    public void Sprite_ShouldCreateASprite_WithGivenWidthAndHeight() {
+        Sprite sprite = Fixtures.Sprite(25, 50);
+        Assert.NotNull(sprite);
+        Assert.AreEqual(25, sprite.texture.width);
+        Assert.AreEqual(50, sprite.texture.height);
+    }
+
+    [Test]
     public void Sprites_ShouldCreateAMatchingNumberOfSprites() {
         Sprite[] sprites = Fixtures.Sprites(2);